+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ## Unreleased CLI
 
+## cli-v0.4.0
+
 * Added `scittlets pack` command to inline `<script>` elements with `src` into the HTML file (#9)
 
 ## v0.3.0